requirements.txt
fastapi
uvicorn
numpy
requests
python-dotenv